+++ src/02_the_Next_Week/07_Volume/07_Volume.cpp
@@ -53,7 +53,7 @@ int main(int argc, char ** argv){
 	Timer timer;
 	timer.Start();
 	Ptr<Operation> imgUpdate = ToPtr(new LambdaOp([&]() {
-		size_t loopMax = glm::max(imgWindow.GetScale(), 1.0);
+		size_t loopMax = static_cast<size_t>(glm::max(imgWindow.GetScale(), 1.0));
 		pixelSet.RandPick(loopMax, pixels);
 
 		int pixelsNum = pixels.size();
@@ -61,7 +61,7 @@ int main(int argc, char ** argv){
 		for (int pixelIdx = 0; pixelIdx < pixelsNum; pixelIdx++) {
 			const uvec2 & pixel = pixels[pixelIdx];
 			rgb color(0);
-			for (size_t k = 0; k < sampleNum; k++) {
+			for (int k = 0; k < sampleNum; k++) {
 				float u = (pixel.x + Math::Rand_F()) / (float)val_ImgWidth;
 				float v = (pixel.y + Math::Rand_F()) / (float)val_ImgHeight;
 				color += RayTracer::Trace(scene->obj, scene->camera->GenRay(u, v));
